Move climate intent to homeassistant integration
