+++ homeassistant/components/climate/__init__.py
@@ -68,7 +68,6 @@ from .const import (  # noqa: F401
     FAN_ON,
     FAN_TOP,
     HVAC_MODES,
-    INTENT_GET_TEMPERATURE,
     INTENT_SET_TEMPERATURE,
     PRESET_ACTIVITY,
     PRESET_AWAY,
